[LMS] Add search functionalities

+++ src/main/java/mk/ukim/finki/lms/dto/EventDTO.java
@@ -13,5 +13,5 @@ public class EventDTO {
   private String eventName;
   private String description;
   private LocalDate eventTime;
-  private Integer numAttendees;
+  private Long numAttendees;
 }
